X.PCA.reduction

Examine state of an assumed complex on a PCA plot using extracted features

Usage

X.PCA.reduction(
  data = NULL,
  average = FALSE,
  plot = FALSE,
  separate.plots = TRUE,
  plot.complexes = NA,
  plot.colors = NA,
  plot.color.by = "condition",
  plot.path = FALSE,
  plot.animate = NA
)

Arguments

data

Data frame with features. Columns are 'protein', 'condition', 'replicate' and feature columns.

average

Logical: Should the replicates be averaged before the PCA? Default is FALSE.

plot

Logical: Should the results be plotted?

separate.plots

Logical: Should the plots be plotted separately for different conditions? Default is FALSE.

plot.complexes

Named list of vectors of complexes.

plot.colors

Vector: Colors to be used for plotting

plot.path

Logical: Should the same proteins by connected between the conditions?

plot.animate

Character string: 'Time' for time-based animation and 'State' for state-based animation. Default is NA for no animation.

plot.colors.by

Character string: By which column should the plots be colored? Default is 'condition'.

Value

List of five elements: $data contains PCA coordinates for each protein, $PCA contains information about the PCA, var.explained shows what percentage of variation is explained by which PC, $plots are PCA plots and $separate.plots is a list of separate PCA plots.

Examples

data.PCAs <- X.PCA.reduction(all.features.data,plot=TRUE,plot.complexes=complexes)
#> Error: object 'all.features.data' not found