X.average.reps

Average predictions for multiple models across replicates.

Usage

X.average.reps(
  data = NULL,
  scores.col = NA,
  weights = NULL,
  evaluate = FALSE,
  standard.set = NULL,
  labels.col = NA,
  eval.metric = "prc"
)

Arguments

data

List of data frame with columns for protein1, protein2 and a column with predictions.

scores.col

Character string: Name of the column with prediction values.

weights

Numeric vector: Typically values between 0-1, giving weights for each replicate for weighted averaging.

evaluate

Logical: Should the averaged predictions be evaluated? Default is FALSE. If TRUE, standard.set must be included

standard.set

Dataframe with columns protein1, protein2 and another column with labels.

labels.col

Character string: Name of the standard.set column with labels.

eval.metric

Character string: Evaluation metric to be used for evaluating the model performance. Options same as in X.evaluate.

Value

A list with four elements: $data is the data.frame with averaged predictions, $eval.metric is the integer specifying the resulting evaluation metric, $eval.data are underlying data for plotting the evaluation metric and $plot is the plot itself.

Examples

final.model <- X.average.reps(data=comp.models.data, evaluate=TRUE,standard.set=GS)
#> Error: object 'comp.models.data' not found