X.clean.meltcurves

Cleans the loaded mass spec data

Usage

X.clean.meltcurves(
  data = NULL,
  keratins = TRUE,
  serum = TRUE,
  trypsin = TRUE,
  average.duplicate.proteins = TRUE,
  remove.nas = TRUE
)

Arguments

data

Data frame with the loaded mass spec data

serum

Logical: Should the serum proteins be removed?

trypsin

Logical: Should trypsin be removed?

average.duplicate.proteins

Logical: Should the duplicate proteins be averaged?

remove.nas

Logical: Should the rows with NA values be removed?

keratin

Logical: Should the keratins be removed?

Value

List of two elements. $data is the data frame with contaminants removed and $removed is the data frame listing the removed contaminants.

Examples

MCdata.clean <- X.clean.meltcurves(MCdata.raw)
#> Error: object 'MCdata.raw' not found