X.network.stats

Evaluate a network of protein complexes

Usage

X.network.stats(
  data = NULL,
  scores.col = NA,
  standard.set = NULL,
  labels.col = NA,
  complex.stats = FALSE,
  interconnected = FALSE
)

Arguments

data

Data frame with columns 'protein1' and 'protein2' and pairwise prediction values.

scores.col

Character string: Name of the data column with model scores.

standard.set

Data frame with columns protein1, protein2 and another column with labels.

labels.col

Character string: Name of the columns with labels.

complex.stats

Logical: Should the statistics specific for protein complexes be calculated? Do not use with unclustered networks. Default is FALSE.

interconnected

Logical: If TRUE, the statistics will be calculated on interconnected network, meaning that even the subunits that are present in the same cluster, but did not pass the prediction cut-off, will be considered interconnected. Default is FALSE.

Value

A list with two elements: $data with underlying data and $plot with the plotted statistics.

Examples

network.stats <- X.network.stats(data)
#> Error in theme_bw(base_size = 12): could not find function "theme_bw"