X.plot.network

Using one of the network building algorithms, cluster initial network into individual protein complexes.

Usage

X.plot.network(
  data = NULL,
  scores.col = NA,
  standard.set = NULL,
  labels.col = NA,
  annotation = NULL,
  design.params = list(),
  export.destination = ".",
  network.name = "PPIN",
  network.collection = "PPINs",
  cytoscape.path = NA,
  cytoscape.waittime = 40
)

Arguments

data

Data frame with pair-wise interactions (columns 'protein1' and 'protein2') and a column with probabilities from the ML model.

scores.col

Character string: Name of the columns with prediction values.

standard.set

Data frame with columns protein1, protein2 and another column with labels.

labels.col

Character string: Name of the columns with labels.

annotation

Data frame with column 'id' with protein IDs and any other annotation columns.

design.params

List of lines of pieces of code to set up the design of the network plot.

export.destination

Folder where the files will be saved. Default is the current directory.

network.name

Title of the network in Cytoscape.

network.collection

Collection of the network in Cytoscape.

cytoscape.path

Path to Cytoscape executable file. If not given, the program will try to find it, which may take a long time.

cytoscape.waittime

Integer: How long should the program wait for the cytoscape to initiate.

Value

Nothing to return, the network is saved in the export.destination in various formats.

Examples

design=list('setNodeShapeDefault("ELLIPSE")','setNodeColorDefault("#5A5A5A")','lockNodeDimensions(TRUE)','setNodeLabelOpacityDefault(1)','setEdgeColorMapping(table.column="complex",list("1","0"), list("#00FF00","#FF0000"),mapping.type="d")','setEdgeLineWidthMapping(mapping.type="c",table.column="pred", widths=c(1,20))')
X.plot.network(final_data,scores.col="score",standard.set=GS,labels.col="complex",annotation=plasmoDB_data%>%rename(protein=Accession), 
              design.params=design, network.name="MAPX_network", cytoscape.path="C://Program Files/Cytoscape_v3.9.1/Cytoscape.exe")
#> Error: object 'final_data' not found