X.postprocess

Using one of the network building algorithms, cluster initial network into individual protein complexes.

Usage

X.postprocess(
  data = NULL,
  mode = "trim",
  scores.col = NA,
  weighted = FALSE,
  zcut = 2,
  init.stats = FALSE,
  final.stats = FALSE,
  standard.set = NULL,
  labels.col = NA
)

Arguments

data

Data frame with pair-wise interactions (columns 'protein1' and 'protein2') and a column with probabilities from ML model.

mode

Character string: 'trim' for trimming of not-well-connected proteins and 'split' for separating of two loosely connected subunits. Default is 'trim'.

scores.col

Character string: Name of the columns with prediction values.

zcut

Numeric: Sub-units with z-score above this value will be removed. Default is 2.

init.stats

Logical: Should the initial network stats be calculated? Default is FALSE.

final.stats

Logical: Should the final network stats be calculated? Default is FALSE.

standard.set

Data frame with columns protein1, protein2 and another column with labels.

labels.col

Character string: Name of the columns with labels.

weight

Logical: Should the prediction values be taken into account for the postprocessing? Default is FALSE

Value

A list with three elements: $data with columns 'protein1', 'protein2' and score contains all data for plotting or further processing of the network, 'stats_initial' contains statistics of the network before clustering. and 'stats_final' contains statistics of the network after clustering.

Examples

network.post.split <- X.postprocess(data=network.refined$data, mode="split", scores.col="score", final.stats=TRUE, 
        standard.set=GS,labels="complex", weighted=FALSE)
#> Error: object 'network.refined' not found

network.post.trim <- X.postprocess(data=network.post.split$data, mode="trim", scores.col="score", final.stats=TRUE, 
        standard.set=GS,labels="complex", weighted=TRUE)
#> Error: object 'network.post.split' not found